RAG-PaperBot/requirements.txt
langchain
langchain-community
langchain-groq
langchain-openai
faiss-cpu
sentence-transformers
pypdf
streamlit
python-dotenv
openai
groq
tiktoken
